+++ lib/redis_client/config.rb
@@ -135,7 +135,7 @@ class RedisClient
 
       super(**kwargs)
 
-      @host = host || uri&.host || DEFAULT_HOST
+      @host = host || uri&.host&.sub(/\A\[(.*)\]\z/, '\1') || DEFAULT_HOST
       @port = port || uri&.port || DEFAULT_PORT
       @path = path
     end